requirements.txt
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.1.1
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.0.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==2.3.7
WTForms==3.2.1
Flask-Migrate==4.0.5
